flask/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
uWSGI==2.0.19.1
Werkzeug==2.0.1
